rename cdn errors
